refactor(status): move benchmark status compaction into bounded read models
