+++ loopx/status.py
@@ -4,9 +4,6 @@ from pathlib import Path
 import re
 from typing import Any
 
-from .benchmarks.read_models.skillsbench_verifier_attribution import (
-    apply_skillsbench_verifier_bootstrap_missing_score_attribution,
-)
 from .control_plane import compact_control_plane_policy
 from .control_plane.status.collection import (
     StatusCollectionContext,
@@ -122,19 +119,9 @@ from .control_plane.runtime.status_classifications import (
     HANDOFF_READY_CLASSIFICATIONS,
     USER_OR_CONTROLLER_CLASSIFICATIONS,
 )
-from .benchmarks.read_models.benchmark_projection import (
-    benchmark_run_source as _benchmark_run_source_read_model,
-    build_benchmark_solution_quality_signals,
-    compact_benchmark_run_core as _compact_benchmark_run_core_read_model,
-    compact_benchmark_run_trials as _compact_benchmark_run_trials_read_model,
-    compact_benchmark_run_validation as _compact_benchmark_run_validation_read_model,
-)
 from .benchmarks.read_models.benchmark_comparison import (
     benchmark_comparison_decision_note as _benchmark_comparison_decision_note_read_model,
     compact_benchmark_comparison as _compact_benchmark_comparison_read_model,
-)
-from .benchmarks.read_models.benchmark_attempt_accounting import (
-    compact_benchmark_attempt_accounting as _compact_benchmark_attempt_accounting,
 )
 from .benchmarks.read_models.benchmark_experiment_report import (
     benchmark_experiment_report_readiness_note,
@@ -145,20 +132,10 @@ from .benchmarks.read_models.benchmark_learning_ledger import (
     compact_benchmark_learning_ledger,
 )
 from .benchmarks.read_models.benchmark_result import compact_benchmark_result
-from .benchmarks.read_models.benchmark_run_execution_contract import (
-    compact_benchmark_run_execution_contract as _compact_benchmark_run_execution_contract,
-)
-from .benchmarks.read_models.benchmark_run_post_execution import (
-    compact_benchmark_run_post_execution_metadata as _compact_benchmark_run_post_execution_metadata,
-    repair_product_mode_lifecycle_missing_attribution as _repair_product_mode_lifecycle_missing_attribution_read_model,
-)
-from .benchmarks.read_models.benchmark_run_pre_execution import (
-    compact_benchmark_run_pre_execution_metadata as _compact_benchmark_run_pre_execution_metadata,
+from .benchmarks.read_models.skillsbench_post_run_debug import (
+    build_skillsbench_post_run_debug_gate,  # noqa: F401
 )
 from .control_plane.runtime.public_safety import (
-    compact_loopx_command_records as _compact_loopx_command_records,
-    compact_numeric_map as _compact_numeric_map,
-    public_safe_compact_list,
     public_safe_compact_text,
 )
 from .control_plane.runtime.active_user_assisted_pilot import (
@@ -194,9 +171,6 @@ from .control_plane.goals.goal_vision import (
 )
 from .control_plane.runtime.session_runtime import (
     compact_session_runtime_projection_from_run,
-)
-from .benchmarks.read_models.skillsbench_post_run_debug import (
-    build_skillsbench_post_run_debug_gate,
 )
 from .control_plane.agents.subagent_activity import (
     MAX_SUBAGENT_ACTIVITY_ITEMS,
@@ -260,6 +234,7 @@ from .control_plane.todos.projection import (
 
 
 _PUBLIC_COMPAT_REEXPORTS = {
+    "build_skillsbench_post_run_debug_gate": "loopx.benchmarks.read_models.skillsbench_post_run_debug",
     "DREAMING_ADVISORY_CLASSIFICATIONS": "loopx.control_plane.runtime.status_classifications",
     "TODO_PROJECTION_DETAIL_POINTER_SCHEMA_VERSION": "loopx.control_plane.work_items.project_asset",
     "TODO_PROJECTION_VIEW_SCHEMA_VERSION": "loopx.control_plane.work_items.project_asset",
@@ -432,1537 +407,6 @@ AUTONOMOUS_RUN_HISTORY_STALL_PATTERN = re.compile(
 )
 
 
-def _compact_benchmark_interaction_counters(value: Any) -> dict[str, Any]:
-    if not isinstance(value, dict):
-        return {}
-
-    compact: dict[str, Any] = {}
-    schema = public_safe_compact_text(value.get("schema_version"), limit=100)
-    if schema:
-        compact["schema_version"] = schema
-    for field in (
-        "prompt_policy_injected",
-        "harness_skill_or_packet_injected",
-        "raw_trace_recorded",
-        "raw_task_prompt_recorded",
-        "controller_trace_present",
-        "loopx_automation_loop",
-        "inner_codex_goal_mode",
-        "curated_skills_visible",
-        "product_mode",
-        "goal_start_product_mode",
-        "verifier_failure_feedback_todo_route",
-        "verifier_failure_feedback_forwarded_to_agent",
-        "verifier_failure_todo_required",
-        "goal_start_plan_observed",
-        "planner_before_todo_write",
-        "same_priority_order_preserved",
-        "selected_todo_claimed",
-        "selected_todo_updated_before_solver",
-        "selected_todo_completed_before_spend",
-        "selected_todo_completed_observed",
-        "non_selected_todos_preserved_open_or_deferred",
-        "quota_spend_missing_after_repeated_complete",
-        "blind_loop",
-        "case_goal_state_packet_present",
-        "case_goal_state_init_required",
-        "case_goal_state_initialized_before_agent",
-        "declared_done_requires_no_remaining_goals",
-        "product_mode_lifecycle_checkpoint_required",
-        "product_mode_solver_activity_required",
-        "product_mode_solver_activity_gap",
-        "product_mode_declared_done_below_passing_reward",
-        "product_mode_no_open_todo_below_passing_reward_stop",
-        "product_mode_typed_repair_required",
-        "product_mode_typed_repair_pending",
-        "product_mode_typed_repair_todo_identity_observed",
-        "product_mode_typed_repair_task_or_validation_delta",
-        "product_mode_typed_repair_delta_observed",
-        "product_mode_typed_repair_terminal",
-        "product_mode_typed_repair_terminal_receipt_consistent",
-        "product_mode_host_local_idle_no_task_output_progress",
-        "product_mode_host_local_idle_no_task_output_progress_stop",
-        "product_mode_final_closeout_superseded_by_official_success",
-        "product_mode_no_tool_call_lifecycle_abort",
-        "agent_declared_done",
-        "agent_declared_no_remaining_goals",
-        "official_feedback_blinded",
-        "reward_feedback_forwarded",
-        "controller_official_feedback_forwarded",
-        "controller_blind_loop",
-        "controller_official_success_observed",
-        "controller_budget_cutoff_before_followup",
-        "benchflow_user_loop_final_verify_recovery_enabled",
-        "benchflow_user_loop_final_verify_recovery_triggered",
-        "benchflow_user_loop_recovery_after_agent_activity",
-        "benchflow_user_loop_recovery_preserved_final_verify",
-        "benchflow_user_loop_recovery_raw_error_recorded",
-        "benchflow_intermediate_soft_verify_final_only",
-        "benchflow_intermediate_soft_verify_raw_output_recorded",
-        "benchflow_intermediate_soft_verify_timeout_enabled",
-        "benchflow_intermediate_soft_verify_timeout_triggered",
-        "benchflow_intermediate_soft_verify_timeout_raw_output_recorded",
-        "benchflow_intermediate_soft_verify_timeout_cleanup_requested",
-        "benchflow_intermediate_soft_verify_timeout_cleanup_raw_logs_read",
-        "benchflow_intermediate_soft_verify_orphan_cleanup_requested",
-        "benchflow_intermediate_soft_verify_orphan_cleanup_raw_logs_read",
-        "private_trajectory_summary_present",
-        "native_goal_worker_route",
-        "native_goal_worker_connected",
-        "native_goal_worker_trace_dir_present",
-        "native_goal_worker_public_trace_read",
-        "native_goal_worker_raw_material_recorded",
-        "remote_command_file_bridge_consumed_by_solver",
-        "remote_command_file_bridge_solver_trace_dir_present",
-        "remote_command_file_bridge_solver_public_trace_read",
-        "remote_command_file_bridge_solver_raw_material_recorded",
-        "remote_command_file_bridge_agent_operation_trace_required",
-        "remote_command_file_bridge_agent_operation_trace_satisfied",
-        "remote_command_file_bridge_driver_lifecycle_trace_present",
-        "remote_command_file_bridge_driver_lifecycle_raw_material_recorded",
-        "host_local_acp_codex_exec_failure_trace_present",
-        "host_local_acp_codex_exec_failure_raw_material_recorded",
-    ):
-        if isinstance(value.get(field), bool):
-            compact[field] = value[field]
-    for field in (
-        "loopx_state_reads",
-        "loopx_state_writes",
-        "loopx_case_state_reads",
-        "loopx_case_state_writes",
-        "heartbeat_count",
-        "controller_action_decisions",
-        "controller_initial_prompt_count",
-        "controller_followup_prompt_count",
-        "controller_stop_decision_count",
-        "controller_reward_observation_count",
-        "controller_round_reward_count",
-        "controller_official_success_observation_count",
-        "controller_first_success_round",
-        "declared_done_round",
-        "planned_todo_count",
-        "planned_p0_count",
-        "agent_todo_complete_unique_todo_count",
-        "selected_todo_complete_count",
-        "selected_todo_duplicate_complete_count",
-        "non_selected_todo_complete_count",
-        "todo_complete_without_todo_id_count",
-        "product_mode_lifecycle_checkpoint_count",
-        "product_mode_lifecycle_checkpoint_round",
-        "product_mode_solver_activity_gap_count",
-        "product_mode_solver_activity_gap_round",
-        "product_mode_declared_done_below_passing_reward_count",
-        "product_mode_declared_done_below_passing_reward_round",
-        "verifier_failure_feedback_todo_prompt_count",
-        "verifier_failure_feedback_todo_round",
-        "open_todo_count",
-        "product_mode_no_open_todo_below_passing_reward_streak",
-        "product_mode_no_open_todo_below_passing_reward_streak_threshold",
-        "product_mode_no_open_todo_below_passing_reward_round",
-        "product_mode_no_open_todo_below_passing_reward_stop_count",
-        "product_mode_no_open_todo_below_passing_reward_stop_round",
-        "product_mode_no_open_todo_below_passing_reward_open_todo_count_public",
-        "product_mode_typed_repair_trigger_round",
-        "product_mode_typed_repair_round_entered",
-        "product_mode_typed_repair_round_entered_count",
-        "product_mode_typed_repair_resolved_round",
-        "product_mode_typed_repair_task_facing_success_delta",
-        "product_mode_typed_repair_terminal_round",
-        "product_mode_typed_repair_open_todo_count_public",
-        "product_mode_host_local_idle_no_task_output_progress_streak",
-        "product_mode_host_local_idle_no_task_output_progress_streak_threshold",
-        "product_mode_host_local_idle_no_task_output_progress_round",
-        "product_mode_host_local_idle_no_task_output_progress_stop_count",
-        "product_mode_host_local_idle_no_task_output_progress_stop_round",
-        "product_mode_host_local_idle_no_task_output_progress_last_failure_trace_count",
-        "product_mode_host_local_idle_no_task_output_progress_acp_tool_calls",
-        "product_mode_host_local_idle_no_task_output_progress_bridge_task_ops",
-        "product_mode_host_local_idle_no_task_output_progress_bridge_task_successes",
-        "product_mode_final_closeout_superseded_round",
-        "product_mode_no_tool_call_lifecycle_abort_count",
-        "product_mode_no_tool_call_lifecycle_abort_round",
-        "controller_verifier_feedback_observation_count",
-        "controller_official_feedback_blinded_count",
-        "controller_official_feedback_forwarded_count",
-        "controller_max_rounds_budget",
-        "benchflow_user_loop_recovery_round",
-        "benchflow_user_loop_recovery_delta_events",
-        "benchflow_user_loop_recovery_delta_tool_calls",
-        "benchflow_intermediate_soft_verify_call_count",
-        "benchflow_intermediate_soft_verify_skipped_count",
-        "benchflow_intermediate_soft_verify_timeout_sec",
-        "benchflow_intermediate_soft_verify_timeout_override_count",
-        "benchflow_intermediate_soft_verify_timeout_cleanup_container_count",
-        "benchflow_intermediate_soft_verify_timeout_cleanup_match_count",
-        "benchflow_intermediate_soft_verify_timeout_cleanup_term_sent_count",
-        "benchflow_intermediate_soft_verify_timeout_cleanup_kill_sent_count",
-        "benchflow_intermediate_soft_verify_timeout_cleanup_alive_after_count",
-        "benchflow_intermediate_soft_verify_orphan_cleanup_container_count",
-        "benchflow_intermediate_soft_verify_orphan_cleanup_match_count",
-        "benchflow_intermediate_soft_verify_orphan_cleanup_term_sent_count",
-        "benchflow_intermediate_soft_verify_orphan_cleanup_kill_sent_count",
-        "benchflow_intermediate_soft_verify_orphan_cleanup_alive_after_count",
-        "private_trajectory_event_count",
-        "private_trajectory_round_count",
-        "private_trajectory_tool_call_count",
-        "loopx_cli_call_count",
-        "loopx_cli_state_read_count",
-        "loopx_cli_state_write_count",
-        "loopx_case_state_path_count",
-        "loopx_case_state_read_count",
-        "loopx_case_state_write_count",
-        "protected_path_mention_count",
-        "protected_path_edit_signal_count",
-        "codex_acp_text_bytes",
-        "append_benchmark_run_success_count",
-        "append_benchmark_run_schema_rejected_count",
-        "worker_counter_trace_trial_count",
-        "worker_benchmark_run_file_count",
-        "worker_benchmark_run_schema_ok_count",
-        "worker_self_validation_official_score_mismatch_count",
-        "worker_validation_scope_ambiguous_official_score_failure_count",
-        "worker_bridge_connected_official_score_failure_count",
-        "worker_startup_blocker_count",
-        "worker_setup_diagnostic_file_count",
-        "worker_setup_diagnostic_schema_ok_count",
-        "worker_submit_eligible_mismatch_count",
-        "worker_bridge_writeback_loss_count",
-        "environment_setup_failure_before_worker_count",
-        "pre_worker_agent_setup_failure_count",
-        "codex_runtime_goal_tool_trial_count",
-        "native_goal_worker_connect_count",
-        "native_goal_worker_trace_count",
-        "native_goal_worker_lifecycle_trace_count",
-        "native_goal_worker_prompt_received_count",
-        "native_goal_worker_ok_count",
-        "native_goal_worker_goal_get_count",
-        "native_goal_worker_turn_start_count",
-        "native_goal_worker_turn_completed_observed_count",
-        "native_goal_worker_assistant_message_present_count",
-        "native_goal_worker_assistant_context_only_count",
-        "native_goal_worker_context_only_recovery_attempted_count",
-        "native_goal_worker_context_only_recovery_succeeded_count",
-        "native_goal_worker_context_only_followup_start_attempted_count",
-        "native_goal_worker_context_only_followup_start_succeeded_count",
-        "native_goal_worker_normal_followup_attempted_count",
-        "native_goal_worker_normal_followup_succeeded_count",
-        "native_goal_worker_normal_followup_start_attempted_count",
-        "native_goal_worker_normal_followup_start_succeeded_count",
-        "native_goal_worker_finish_guard_followup_attempted_count",
-        "native_goal_worker_finish_guard_followup_succeeded_count",
-        "native_goal_worker_finish_guard_followup_start_attempted_count",
-        "native_goal_worker_finish_guard_followup_start_succeeded_count",
-        "native_goal_worker_incomplete_turn_status_count",
-        "native_goal_worker_incomplete_after_completion_event_count",
-        "native_goal_worker_transport_reconnect_attempted_count",
-        "native_goal_worker_transport_reconnect_succeeded_count",
-        "native_goal_worker_goal_reactivation_attempted_count",
-        "native_goal_worker_goal_reactivation_succeeded_count",
-        "native_goal_worker_post_context_assistant_chars_total",
-        "native_goal_worker_first_action_observed_count",
-        "native_goal_worker_effective_action_observed_count",
-        "remote_command_file_bridge_solver_trace_count",
-        "remote_command_file_bridge_solver_probe_ready_count",
-        "remote_command_file_bridge_solver_operation_count",
-        "remote_command_file_bridge_agent_operation_trace_count",
-        "remote_command_file_bridge_agent_request_count",
-        "remote_command_file_bridge_agent_success_count",
-        "remote_command_file_bridge_agent_failure_count",
-        "remote_command_file_bridge_agent_loopx_cli_call_count",
-        "remote_command_file_bridge_agent_loopx_state_read_count",
-        "remote_command_file_bridge_agent_loopx_state_write_count",
-        "remote_command_file_bridge_agent_todo_closeout_count",
-        "remote_command_file_bridge_agent_refresh_state_count",
-        "remote_command_file_bridge_agent_quota_spend_slot_count",
-        "remote_command_file_bridge_agent_task_facing_operation_count",
-        "remote_command_file_bridge_agent_task_facing_success_count",
-        "remote_command_file_bridge_agent_task_facing_failure_count",
-        "remote_command_file_bridge_driver_lifecycle_trace_count",
-        "remote_command_file_bridge_driver_lifecycle_checkpoint_count",
-        "remote_command_file_bridge_driver_lifecycle_request_count",
-        "remote_command_file_bridge_driver_lifecycle_success_count",
-        "remote_command_file_bridge_driver_lifecycle_failure_count",
-        "remote_command_file_bridge_driver_lifecycle_loopx_cli_call_count",
-        "remote_command_file_bridge_driver_lifecycle_loopx_state_read_count",
-        "remote_command_file_bridge_driver_lifecycle_loopx_state_write_count",
-        "host_local_acp_codex_exec_failure_trace_count",
-        "host_local_acp_codex_exec_recoverable_failure_trace_count",
-        "host_local_acp_codex_exec_fatal_failure_trace_count",
-    ):
-        if isinstance(value.get(field), int) and not isinstance(value.get(field), bool):
-            compact[field] = value[field]
-    for field in (
-        "product_mode_declared_done_below_passing_reward_score",
-        "product_mode_no_open_todo_below_passing_reward_score",
-        "product_mode_host_local_idle_no_task_output_progress_score",
-    ):
-        raw = value.get(field)
-        if isinstance(raw, (int, float)) and not isinstance(raw, bool):
-            compact[field] = float(raw)
-    for field in (
-        "case_result_writeback",
-        "counter_trust_level",
-        "controller_trace_schema_version",
-        "controller_trace_publicness",
-        "case_goal_state_init_status",
-        "case_goal_state_init_failed_phase",
-        "case_goal_state_schema_version",
-        "product_mode_lifecycle_checkpoint_missing_reason",
-        "product_mode_solver_activity_missing_reason",
-        "product_mode_declared_done_below_passing_reward_score_status",
-        "product_mode_no_open_todo_below_passing_reward_score_status",
-        "product_mode_host_local_idle_no_task_output_progress_score_status",
-        "product_mode_host_local_idle_no_task_output_progress_category",
-        "product_mode_host_local_idle_no_task_output_progress_policy",
-        "product_mode_declared_done_policy",
-        "product_mode_typed_repair_policy_id",
-        "product_mode_typed_repair_terminal_reason",
-        "product_mode_final_closeout_superseded_reason",
-        "controller_budget_cutoff_reason",
-        "benchflow_user_loop_recovery_stage",
-        "benchflow_user_loop_recovery_exception_type",
-        "benchflow_intermediate_soft_verify_policy",
-        "benchflow_intermediate_soft_verify_timeout_stage",
-        "benchflow_intermediate_soft_verify_timeout_cleanup_status",
-        "benchflow_intermediate_soft_verify_orphan_cleanup_status",
-        "remote_command_file_bridge_agent_operation_trace_status",
-        "remote_command_file_bridge_consumption_decision",
-        "remote_command_file_bridge_driver_lifecycle_execution_style",
-        "native_goal_worker_reasoning_effort",
-        "host_local_acp_codex_exec_failure_category",
-        "host_local_acp_bridge_progress_status",
-        "host_local_acp_bridge_progress_signal_source",
-        "last_decision",
-        "worker_submit_eligible_mismatch_reason",
-        "worker_bridge_writeback_loss_reason",
-    ):
-        text = public_safe_compact_text(value.get(field), limit=100)
-        if text:
-            compact[field] = text
-    case_state_path = public_safe_compact_text(
-        value.get("case_goal_state_path"),
-        limit=180,
-    )
-    if (
-        case_state_path
-        and "/.codex/goals/" in case_state_path
-        and case_state_path.endswith("/ACTIVE_GOAL_STATE.md")
-        and not re.search(r"^/(Users|private|var/folders)/", case_state_path)
-    ):
-        compact["case_goal_state_path"] = case_state_path
-
-    for field in (
-        "codex_runtime_goal_tool_calls",
-        "trajectory_action_category_counts",
-        "loopx_cli_state_usage_counts",
-        "remote_command_file_bridge_agent_returncode_counts",
-        "remote_command_file_bridge_agent_loopx_subcommand_counts",
-        "remote_command_file_bridge_agent_successful_loopx_subcommand_counts",
-        "remote_command_file_bridge_driver_lifecycle_command_counts",
-        "remote_command_file_bridge_driver_lifecycle_returncode_counts",
-    ):
-        calls = _compact_numeric_map(value.get(field))
-        if calls:
-            compact[field] = calls
-    selected_p0_todo_id = public_safe_compact_text(
-        value.get("selected_p0_todo_id"),
-        limit=100,
-    )
-    if selected_p0_todo_id:
-        compact["selected_p0_todo_id"] = selected_p0_todo_id
-    planned_todo_ids = public_safe_compact_list(
-        value.get("planned_todo_ids"),
-        limit=8,
-    )
-    if planned_todo_ids:
-        compact["planned_todo_ids"] = planned_todo_ids
-    planned_todo_texts = public_safe_compact_list(
-        value.get("planned_todo_texts_public_safe"),
-        limit=8,
-    )
-    if planned_todo_texts:
-        compact["planned_todo_texts_public_safe"] = planned_todo_texts
-    command_records = _compact_loopx_command_records(
-        value.get("remote_command_file_bridge_agent_successful_loopx_command_records")
-    )
-    if command_records:
-        compact[
-            "remote_command_file_bridge_agent_successful_loopx_command_records"
-        ] = command_records
-    raw_loopx_cli_calls = value.get("loopx_cli_calls")
-    if isinstance(raw_loopx_cli_calls, dict):
-        calls = _compact_numeric_map(raw_loopx_cli_calls)
-        if calls:
-            compact["loopx_cli_calls"] = calls
-    elif isinstance(raw_loopx_cli_calls, list):
-        calls: list[dict[str, Any]] = []
-        for item in raw_loopx_cli_calls[:8]:
-            if not isinstance(item, dict):
-                continue
-            call: dict[str, Any] = {}
-            round_value = item.get("round")
-            if (
-                isinstance(round_value, int)
-                and not isinstance(round_value, bool)
-                and round_value > 0
-            ):
-                call["round"] = round_value
-            command = public_safe_compact_text(item.get("command"), limit=120)
-            if command:
-                call["command"] = command
-            flags = item.get("flags")
-            if isinstance(flags, list):
-                compact_flags = [
-                    flag
-                    for flag in (
-                        public_safe_compact_text(flag, limit=60)
-                        for flag in flags[:8]
-                    )
-                    if flag
-                ]
-                if compact_flags:
-                    call["flags"] = compact_flags
-            if isinstance(item.get("raw_title_copied"), bool):
-                call["raw_title_copied"] = item["raw_title_copied"]
-            if isinstance(item.get("raw_output_copied"), bool):
-                call["raw_output_copied"] = item["raw_output_copied"]
-            if call:
-                calls.append(call)
-        if calls:
-            compact["loopx_cli_calls"] = calls
-
-    return compact
-
-
-def _compact_benchmark_preflight_guard(value: Any) -> dict[str, Any]:
-    if not isinstance(value, dict):
-        return {}
-
-    compact: dict[str, Any] = {}
-    for field in (
-        "schema_version",
-        "first_blocker",
-        "loopx_mode_kwarg",
-        "codex_goal_mode_invocation_surface",
-        "codex_goal_mode_required_invocation_surface",
-        "codex_goal_mode_baseline_claim_blocker",
-        "codex_app_server_goal_worker_plan_schema",
-        "runner_binary_resolution_policy",
-        "simulator_setting",
-    ):
-        text = public_safe_compact_text(value.get(field), limit=120)
-        if text:
-            compact[field] = text
-    for field in (
-        "runner_surface_checked",
-        "local_execution_surface_checked",
-        "codex_cli_surface_checked",
-        "auth_surface_names_only",
-        "auth_values_read",
-        "artifact_redaction_required",
-        "task_material_ready_required",
-        "access_packet_prompt_injection_checked",
-        "trace_counter_extraction_contract_checked",
-        "loopx_mode_kwarg_checked",
-        "codex_goal_mode_invocation_surface_checked",
-        "codex_app_server_goal_baseline_requested",
-        "codex_app_server_goal_worker_adapter_present",
-        "codex_app_server_goal_worker_adapter_absent",
-        "codex_app_server_goal_worker_turn_start_required",
-        "codex_app_server_goal_proof_present",
-        "codex_goal_mode_baseline_claim_allowed",
-        "loopx_access_packet_absent",
-        "loopx_cli_bridge_absent",
-        "active_cli_bridge_enabled",
-        "claim_requires_worker_cli_calls",
-        "real_interface_use_observed",
-        "uplift_claim_allowed",
-        "active_user_assisted_treatment",
-        "simulator_to_worker_injection_channel_available",
-        "interactive_user_message_injection_checked",
-        "initial_prompt_only_is_not_active_intervention",
-        "no_oracle_audit_required",
-        "assisted_score_kept_separate_from_official",
-        "uvx_cli_present",
-        "uvx_version_probe_ok",
-        "docker_cli_present",
-        "docker_version_probe_ok",
-        "docker_server_available",
-        "colima_cli_present",
-        "colima_status_probe_ok",
-        "codex_cli_present",
-        "codex_version_probe_ok",
-    ):
-        if isinstance(value.get(field), bool):
-            compact[field] = value[field]
-    text = public_safe_compact_text(value.get("worker_cli_bridge_surface"), limit=120)
-    if text:
-        compact["worker_cli_bridge_surface"] = text
-    for field in ("required_worker_loopx_cli_call_total_min",):
-        if isinstance(value.get(field), int) and not isinstance(value.get(field), bool):
-            compact[field] = value[field]
-    return compact
-
-
-def _compact_benchmark_compose_setup_diagnostic(value: Any) -> dict[str, Any]:
-    if not isinstance(value, dict):
-        return {}
-
-    compact: dict[str, Any] = {}
-    for field in (
-        "schema_version",
-        "status",
-        "route",
-        "failure_class",
-        "runner_prerequisite_status",
-        "task_setup_preflight_status",
-        "fingerprint_confidence",
-        "runner_error_len_bucket",
-        "primary_setup_failure_category",
-        "apt_failure_subtype",
-        "pip_failure_subtype",
-        "retryability",
-        "next_diagnostic_action",
-    ):
-        text = public_safe_compact_text(value.get(field), limit=180)
-        if text:
-            compact[field] = text
-    for field in (
-        "compose_setup_failure",
-        "unclassified_compose_failure",
-        "docker_daemon_unavailable",
-        "apt_repository_failure",
-        "pip_bootstrap_failure",
-        "volume_mount_failure",
-        "environment_setup_failure",
-        "agent_rounds_started",
-        "official_score_missing",
-        "official_result_json_materialized",
-        "case_attempt_budget_should_count",
-        "runner_launch_preflight_passed",
-        "apt_setup_risk_detected",
-        "apt_retry_patch_required",
-        "verifier_uv_bootstrap_risk_detected",
-        "verifier_uv_bootstrap_mirror_patch_required",
-        "verifier_uv_bootstrap_mirror_patch_applied",
-        "staged_task_prepared",
-        "task_skills_removed",
-        "codex_acp_runtime_tools_patch_applied",
-        "resource_cap_applied",
-        "raw_error_recorded",
-        "raw_logs_read",
-        "raw_task_text_read",
-        "raw_trajectory_read",
-    ):
-        if isinstance(value.get(field), bool):
-            compact[field] = value[field]
-    for field in (
-        "heartbeat_count",
-        "controller_action_decision_count",
-        "trajectory_round_count",
-        "trajectory_tool_call_count",
-        "loopx_cli_call_count",
-        "round_reward_count",
-        "setup_stall_timeout_requested_sec",
-        "setup_stall_timeout_sec",
-        "progress_completed_trials",
-        "progress_errored_trials",
-    ):
-        if isinstance(value.get(field), int) and not isinstance(value.get(field), bool):
-            compact[field] = value[field]
-    patterns = public_safe_compact_list(
-        value.get("fingerprint_matched_patterns"),
-        limit=MAX_BENCHMARK_RUN_LIST_ITEMS,
-    )
-    if patterns:
-        compact["fingerprint_matched_patterns"] = patterns
-    for field in (
-        "terminal_failure_dependency_classes",
-        "terminal_failure_reason_codes",
-        "terminal_failure_dependency_endpoints",
-    ):
-        values = public_safe_compact_list(
-            value.get(field),
-            limit=MAX_BENCHMARK_RUN_LIST_ITEMS * 2,
-        )
-        if values:
-            compact[field] = values
-    return compact
-
-
-_SKILLSBENCH_PRE_AGENT_SETUP_STATUS_LABELS = {
-    "compose_setup_blocked_before_agent_rounds": (
-        "skillsbench_compose_setup_blocked_before_agent_rounds"
-    ),
-    "runner_setup_blocked_before_agent_rounds": (
-        "skillsbench_runner_setup_blocked_before_agent_rounds"
-    ),
-}
-
-
-def _skillsbench_compact_official_score_missing(compact: dict[str, Any]) -> bool:
-    official = (
-        compact.get("official_task_score")
-        if isinstance(compact.get("official_task_score"), dict)
-        else {}
-    )
-    value = official.get("value")
-    if isinstance(value, (int, float)) and not isinstance(value, bool):
-        return False
-    value = compact.get("official_score")
-    return not (isinstance(value, (int, float)) and not isinstance(value, bool))
-
-
-def _sync_skillsbench_runner_failure_root_blockers(
-    compact: dict[str, Any],
-) -> None:
-    runner_failure = compact.get("runner_failure")
-    if not isinstance(runner_failure, dict):
-        return
-    if not _skillsbench_compact_official_score_missing(compact):
-        return
-
-    blocker = public_safe_compact_text(
-        compact.get("score_failure_attribution"),
-        limit=140,
-    )
-    if blocker in {None, "none", "score_missing"}:
-        blocker = public_safe_compact_text(
-            runner_failure.get("failure_class"),
-            limit=140,
-        )
-    if not blocker:
-        return
-
-    replaceable = {
-        None,
-        "none",
-        "score_missing",
-        "skillsbench_runner_error",
-    }
-    for field in ("first_blocker", "repeat_blocked_by"):
-        if compact.get(field) in replaceable:
-            compact[field] = blocker
-
-
-def _skillsbench_compact_pre_agent_setup_label(compact: dict[str, Any]) -> str:
-    diagnostic = compact.get("compose_setup_diagnostic")
-    if not isinstance(diagnostic, dict):
-        return ""
-    label = _SKILLSBENCH_PRE_AGENT_SETUP_STATUS_LABELS.get(
-        str(diagnostic.get("status") or "")
-    )
-    if not label:
-        return ""
-    if compact.get("mode") != "skillsbench_codex_app_server_goal_baseline":
-        return ""
-    validation = (
-        compact.get("validation") if isinstance(compact.get("validation"), dict) else {}
-    )
-    native_route = (
-        compact.get("native_goal_worker_route")
-        if "native_goal_worker_route" in compact
-        else validation.get("native_goal_worker_route")
-    )
-    if native_route is not True:
-        return ""
-    native_connected = (
-        compact.get("native_goal_worker_connected")
-        if "native_goal_worker_connected" in compact
-        else validation.get("native_goal_worker_connected")
-    )
-    if native_connected is True:
-        return ""
-    trace_count = (
-        compact.get("native_goal_worker_trace_count")
-        if "native_goal_worker_trace_count" in compact
-        else validation.get("native_goal_worker_trace_count")
-    )
-    if (
-        isinstance(trace_count, int)
-        and not isinstance(trace_count, bool)
-        and trace_count > 0
-    ):
-        return ""
-    if diagnostic.get("agent_rounds_started") is True:
-        return ""
-    if not _skillsbench_compact_official_score_missing(compact):
-        return ""
-    return label
-
-
-def _apply_skillsbench_pre_agent_setup_compact_projection(
-    compact: dict[str, Any],
-) -> None:
-    label = _skillsbench_compact_pre_agent_setup_label(compact)
-    if not label:
-        return
-    current = str(compact.get("score_failure_attribution") or "")
-    if (
-        current in {"", "none", "score_missing", "skillsbench_runner_error"}
-        or current.startswith("skillsbench_native_goal_worker_")
-    ):
-        compact["score_failure_attribution"] = label
-        compact["first_blocker"] = label
-    labels = [
-        item
-        for item in compact.get("failure_attribution_labels", [])
-        if isinstance(item, str) and item
-    ]
-    for item in (
-        label,
-        "skillsbench_app_server_goal_pre_agent_materialization_blocked",
-        "skillsbench_runner_setup_error",
-    ):
-        if item not in labels:
-            labels.append(item)
-    compact["failure_attribution_labels"] = labels[:MAX_BENCHMARK_RUN_LIST_ITEMS]
-    attempt_accounting = compact.get("attempt_accounting")
-    if isinstance(attempt_accounting, dict):
-        attempt_accounting["failure_label"] = label
-        attempt_accounting["failure_class"] = "job_materialization_failed"
-    runner_failure = compact.get("runner_failure")
-    if isinstance(runner_failure, dict):
-        runner_failure["exception_type"] = label
-        runner_failure["failure_class"] = label
-        runner_failure["pre_agent_setup_materialization_blocked"] = True
-    validation = compact.get("validation")
-    if isinstance(validation, dict):
-        failed = [
-            item
-            for item in validation.get("failed_checks", [])
-            if isinstance(item, str) and item
-        ]
-        failed = [
-            item for item in failed if item != "native_goal_worker_public_trace_missing"
-        ]
-        if "pre_agent_setup_materialization_blocked" not in failed:
-            failed.append("pre_agent_setup_materialization_blocked")
-        validation["failed_checks"] = failed[:MAX_BENCHMARK_RUN_LIST_ITEMS]
-        validation["all_passed"] = False
-    compact["pre_agent_setup_materialization_blocked"] = True
-    compact["native_goal_worker_pre_agent_setup_blocked"] = True
-
-
-def _mark_skillsbench_pre_task_attempt_accounting(
-    compact: dict[str, Any],
-    *,
-    failure_label: str,
-) -> None:
-    attempt_accounting = compact.get("attempt_accounting")
-    if not isinstance(attempt_accounting, dict):
-        return
-    attempt_accounting["lifecycle_phase"] = "runner_accepted_args"
-    attempt_accounting["failure_label"] = failure_label
-    attempt_accounting["failure_class"] = "job_materialization_failed"
-    attempt_accounting["launcher_attempt_countable"] = True
-    for field in (
-        "case_attempt_countable",
-        "solver_attempt_countable",
-        "verifier_attempt_countable",
-        "official_score_attempt_countable",
-    ):
-        attempt_accounting[field] = False
-    attempts = attempt_accounting.get("attempts")
-    if not isinstance(attempts, dict):
-        return
-    launcher = attempts.get("launcher")
-    if isinstance(launcher, dict):
-        launcher["attempted"] = True
-        launcher["countable"] = True
-    for phase_name in ("case", "solver", "verifier", "official_score"):
-        phase = attempts.get(phase_name)
-        if isinstance(phase, dict):
-            phase["attempted"] = False
-            phase["countable"] = False
-
-
-def _apply_skillsbench_benchmark_egress_preflight_compact_projection(
-    compact: dict[str, Any],
-    *,
-    source: dict[str, Any] | None = None,
-) -> None:
-    source = source if isinstance(source, dict) else {}
-    source_runner_config = (
-        source.get("runner_config")
-        if isinstance(source.get("runner_config"), dict)
-        else {}
-    )
-    proxy_required = (
-        compact.get("benchmark_egress_proxy_required") is True
-        or source_runner_config.get("benchmark_egress_proxy_required") is True
-    )
-    proxy_ready = (
-        compact.get("benchmark_egress_proxy_ready") is True
-        or source_runner_config.get("benchmark_egress_proxy_ready") is True
-    )
-    proxy_status = public_safe_compact_text(
-        compact.get("benchmark_egress_proxy_status")
-        or source_runner_config.get("benchmark_egress_proxy_status"),
-        limit=120,
-    )
-    benchmark_failure_statuses = {
-        "failed",
-        "invalid_proxy_value",
-        "missing_required_proxy",
-        "proxy_auth_required",
-        "proxy_connect_rejected",
-        "unsupported_proxy_scheme",
-    }
-    benchmark_blocked = bool(
-        proxy_required
-        and not proxy_ready
-        and proxy_status in benchmark_failure_statuses
-    )
-    codex_required = (
-        compact.get("codex_api_egress_preflight_required") is True
-        or source_runner_config.get("codex_api_egress_preflight_required") is True
-    )
-    codex_ready = (
-        compact.get("codex_api_egress_preflight_ready") is True
-        or source_runner_config.get("codex_api_egress_preflight_ready") is True
-    )
-    codex_status = public_safe_compact_text(
-        compact.get("codex_api_egress_preflight_status")
-        or source_runner_config.get("codex_api_egress_preflight_status"),
-        limit=120,
-    )
-    codex_failure_statuses = {
-        "failed",
-        "missing_reverse_tunnel_proxy",
-        "proxy_auth_required",
-        "proxy_connect_rejected",
-        "unsupported_egress_mode",
-        "unsupported_proxy_scheme",
-    }
-    codex_blocked = bool(
-        codex_required and not codex_ready and codex_status in codex_failure_statuses
-    )
-    if not benchmark_blocked and not codex_blocked:
-        return
-    if not _skillsbench_compact_official_score_missing(compact):
-        return
-
-    if benchmark_blocked:
-        label = "skillsbench_benchmark_egress_proxy_preflight_blocked"
-        status_label = f"skillsbench_benchmark_egress_proxy_{proxy_status}"
-        blocker_key = "benchmark_egress_proxy_preflight_blocked"
-    else:
-        label = "skillsbench_codex_api_egress_preflight_blocked"
-        status_label = f"skillsbench_codex_api_egress_{codex_status}"
-        blocker_key = "codex_api_egress_preflight_blocked"
-    compact["score_failure_attribution"] = label
-    compact["first_blocker"] = label
-    compact["repeat_blocked_by"] = label
-    compact["official_score_comparable_to_native_codex"] = False
-    compact["official_score_comparable_to_loopx_treatment"] = False
-
-    labels = [
-        item
-        for item in compact.get("failure_attribution_labels", [])
-        if isinstance(item, str)
-        and item
-        and item
-        not in {
-            "skillsbench_product_mode_uncountable_treatment",
-            "skillsbench_remote_bridge_agent_operation_trace_missing",
-            "skillsbench_verifier_package_install_risk",
-            "skillsbench_verifier_uv_install_or_download_failure",
-            "verifier_dependency_install_failure",
-        }
-    ]
-    primary_labels = (
-        label,
-        "skillsbench_environment_setup_error",
-        status_label,
-    )
-    labels = list(primary_labels) + [
-        item for item in labels if item not in primary_labels
-    ]
-    compact["failure_attribution_labels"] = labels[:MAX_BENCHMARK_RUN_LIST_ITEMS]
-
-    _mark_skillsbench_pre_task_attempt_accounting(
-        compact,
-        failure_label=label,
-    )
-    runner_failure = compact.get("runner_failure")
-    if isinstance(runner_failure, dict):
-        runner_failure["failure_class"] = label
-        runner_failure[blocker_key] = True
-
-    validation = (
-        compact.get("validation")
-        if isinstance(compact.get("validation"), dict)
-        else {}
-    )
-    validation[blocker_key] = True
-    validation["raw_verifier_output_read"] = False
-    validation["all_passed"] = False
-    compact["validation"] = validation
-
-    if benchmark_blocked:
-        compact["benchmark_egress_proxy_diagnostic"] = {
-            "schema_version": "skillsbench_benchmark_egress_proxy_diagnostic_v0",
-            "status": "benchmark_egress_proxy_preflight_blocked",
-            "score_failure_attribution": label,
-            "proxy_required": True,
-            "proxy_ready": False,
-            "proxy_status": proxy_status,
-            "proxy_error_kind": public_safe_compact_text(
-                compact.get("benchmark_egress_proxy_error_kind")
-                or source_runner_config.get("benchmark_egress_proxy_error_kind"),
-                limit=120,
-            ),
-            "proxy_mode_requested": public_safe_compact_text(
-                compact.get("benchmark_egress_proxy_mode_requested")
-                or source_runner_config.get("benchmark_egress_proxy_mode_requested"),
-                limit=80,
-            ),
-            "proxy_mode_effective": public_safe_compact_text(
-                compact.get("benchmark_egress_proxy_mode_effective")
-                or source_runner_config.get("benchmark_egress_proxy_mode_effective"),
-                limit=80,
-            ),
-            "proxy_url_recorded": False,
-            "raw_logs_read": False,
-            "raw_task_text_read": False,
-            "raw_trajectory_read": False,
-            "next_diagnostic_action": (
-                "configure_valid_private_benchmark_egress_proxy"
-            ),
-        }
-        return
-
-    compact["codex_api_egress_diagnostic"] = {
-        "schema_version": "skillsbench_codex_api_egress_diagnostic_v0",
-        "status": "codex_api_egress_preflight_blocked",
-        "score_failure_attribution": label,
-        "egress_required": True,
-        "egress_ready": False,
-        "egress_status": codex_status,
-        "egress_error_kind": public_safe_compact_text(
-            compact.get("codex_api_egress_preflight_error_kind")
-            or source_runner_config.get("codex_api_egress_preflight_error_kind"),
-            limit=120,
-        ),
-        "egress_mode_requested": public_safe_compact_text(
-            compact.get("codex_api_egress_mode_requested")
-            or source_runner_config.get("codex_api_egress_mode_requested"),
-            limit=80,
-        ),
-        "egress_mode_effective": public_safe_compact_text(
-            compact.get("codex_api_egress_mode_resolved")
-            or source_runner_config.get("codex_api_egress_mode_resolved"),
-            limit=80,
-        ),
-        "reverse_tunnel_required": (
-            compact.get("codex_api_reverse_tunnel_required") is True
-            or source_runner_config.get("codex_api_reverse_tunnel_required") is True
-        ),
-        "proxy_url_recorded": False,
-        "raw_logs_read": False,
-        "raw_task_text_read": False,
-        "raw_trajectory_read": False,
-        "next_diagnostic_action": "restore_required_private_egress_then_retry",
-    }
-
-
-def _apply_skillsbench_runner_source_fingerprint_compact_projection(
-    compact: dict[str, Any],
-    *,
-    source: dict[str, Any] | None = None,
-) -> None:
-    source = source if isinstance(source, dict) else {}
-    prerequisites = (
-        compact.get("runner_prerequisites")
-        if isinstance(compact.get("runner_prerequisites"), dict)
-        else {}
-    )
-    source_config = (
-        source.get("runner_config")
-        if isinstance(source.get("runner_config"), dict)
-        else {}
-    )
-    status = public_safe_compact_text(
-        prerequisites.get("loopx_runner_source_fingerprint_status")
-        or source_config.get("loopx_runner_source_fingerprint_status"),
-        limit=80,
-    )
-    blocker = public_safe_compact_text(
-        prerequisites.get("loopx_runner_source_first_blocker")
-        or source_config.get("loopx_runner_source_first_blocker"),
-        limit=120,
-    )
-    if (
-        status != "mismatched_expected"
-        or blocker != "loopx_runner_source_git_head_mismatch"
-        or not _skillsbench_compact_official_score_missing(compact)
-    ):
-        return
-
-    compact["score_failure_attribution"] = blocker
-    compact["first_blocker"] = blocker
-    compact["repeat_blocked_by"] = blocker
-    compact["official_score_comparable_to_native_codex"] = False
-    compact["official_score_comparable_to_loopx_treatment"] = False
-
-    verifier_labels = {
-        "verifier_dependency_bootstrap_timeout",
-        "verifier_dependency_install_failure",
-        "verifier_uv_install_or_download_failure",
-        "skillsbench_verifier_bootstrap_missing_official_score",
-        "skillsbench_verifier_bootstrap_preflight_blocked",
-        "skillsbench_verifier_package_install_risk",
-    }
-    labels = [
-        item
-        for item in compact.get("failure_attribution_labels", [])
-        if isinstance(item, str) and item and item not in verifier_labels
-    ]
-    primary_labels = (
-        blocker,
-        "skillsbench_runner_source_fingerprint_mismatch",
-        "skillsbench_runner_setup_error",
-    )
-    labels = list(primary_labels) + [
-        item for item in labels if item not in primary_labels
-    ]
-    compact["failure_attribution_labels"] = labels[:MAX_BENCHMARK_RUN_LIST_ITEMS]
-
-    _mark_skillsbench_pre_task_attempt_accounting(
-        compact,
-        failure_label=blocker,
-    )
-    runner_failure = compact.get("runner_failure")
-    if isinstance(runner_failure, dict):
-        runner_failure["failure_class"] = blocker
-        runner_failure["runner_source_fingerprint_mismatch"] = True
-
-    validation = (
-        compact.get("validation")
-        if isinstance(compact.get("validation"), dict)
-        else {}
-    )
-    validation["runner_source_fingerprint_mismatch"] = True
-    validation["all_passed"] = False
-    compact["validation"] = validation
-
-    compose_setup_diagnostic = compact.get("compose_setup_diagnostic")
-    if isinstance(compose_setup_diagnostic, dict):
-        compose_setup_diagnostic["case_attempt_budget_should_count"] = False
-
-
-def _compact_benchmark_private_runner_launch(value: Any) -> dict[str, Any]:
-    if not isinstance(value, dict):
-        return {}
-
-    compact: dict[str, Any] = {}
-    for field in (
-        "schema_version",
-        "launch_schema_version",
-        "first_blocker",
-        "argv_binary_name",
-        "codex_goal_mode_invocation_surface",
-        "codex_goal_mode_required_invocation_surface",
-        "codex_goal_mode_baseline_claim_blocker",
-        "codex_app_server_goal_worker_plan_schema",
-        "task_material_readiness_status",
-        "task_material_first_blocker",
-    ):
-        text = public_safe_compact_text(value.get(field), limit=120)
-        if text:
-            compact[field] = text
-    for field in (
-        "uses_private_runner_env",
-        "ready",
-        "argv_present",
-        "argv_binary_resolved_for_private_launch",
-        "no_upload_boundary",
-        "submit_eligible",
-        "env_path_present",
-        "active_user_writable_mount_requested",
-        "active_user_writable_mount_target_present",
-        "agent_import_path_present",
-        "loopx_agent_kwargs_present",
-        "codex_goal_mode_baseline_requested",
-        "codex_app_server_goal_baseline_requested",
-        "codex_app_server_goal_worker_adapter_present",
-        "codex_app_server_goal_worker_turn_start_required",
-        "codex_app_server_goal_proof_present",
-        "codex_goal_mode_baseline_claim_allowed",
-        "loopx_access_packet_absent",
-        "loopx_worker_bridge_requested",
-        "worker_materialization_probe_only",
-        "task_material_readiness_checked",
-        "task_material_ready_required",
-        "task_material_ready",
-        "setup_timeout_repair_profile",
-        "auth_values_recorded",
-        "raw_env_recorded",
-        "raw_paths_recorded",
-    ):
-        if isinstance(value.get(field), bool):
-            compact[field] = value[field]
-    for field in (
-        "env_probe_path_coverage_count",
-        "task_material_candidate_count",
-        "task_material_instruction_md_present_count",
-        "task_material_task_toml_present_count",
-    ):
-        count = value.get(field)
-        if isinstance(count, int) and not isinstance(count, bool):
-            compact[field] = count
-    active_user_mount_count = value.get("active_user_writable_mount_count")
-    if isinstance(active_user_mount_count, int) and not isinstance(active_user_mount_count, bool):
-        compact["active_user_writable_mount_count"] = active_user_mount_count
-    coverage = value.get("env_probe_path_coverage")
-    compact_coverage: dict[str, bool] = {}
-    if isinstance(coverage, dict):
-        compact_coverage = {
-            str(key): ready
-            for key, ready in coverage.items()
-            if isinstance(key, str) and isinstance(ready, bool)
-        }
-    if compact_coverage:
-        compact["env_probe_path_coverage"] = compact_coverage
-    policy = value.get("timeout_multiplier_policy")
-    if isinstance(policy, dict):
-        compact_policy: dict[str, Any] = {}
-        schema = public_safe_compact_text(policy.get("schema_version"), limit=120)
-        if schema:
-            compact_policy["schema_version"] = schema
-        for field in (
-            "any_timeout_multiplier_present",
-            "non_default_timeout_multiplier_present",
-            "agent_setup_timeout_multiplier_present",
-            "changes_official_benchmark_timeout",
-            "leaderboard_claim_allowed",
-            "raw_argv_recorded",
-        ):
-            if isinstance(policy.get(field), bool):
-                compact_policy[field] = policy[field]
-        multipliers = policy.get("multipliers")
-        if isinstance(multipliers, dict):
-            compact_multipliers = {
-                str(key): value
-                for key, value in multipliers.items()
-                if isinstance(key, str)
-                and isinstance(value, (int, float))
-                and not isinstance(value, bool)
-            }
-            if compact_multipliers:
-                compact_policy["multipliers"] = compact_multipliers
-        if compact_policy:
-            compact["timeout_multiplier_policy"] = compact_policy
-    repair_profile = _compact_benchmark_repair_profile(value.get("repair_profile"))
-    if repair_profile:
-        compact["repair_profile"] = repair_profile
-    readiness = _compact_agent_setup_readiness(value.get("agent_setup_readiness"))
-    if readiness:
-        compact["agent_setup_readiness"] = readiness
-    names = public_safe_compact_list(
-        value.get("auth_surface_names_present"),
-        limit=MAX_BENCHMARK_RUN_LIST_ITEMS,
-    )
-    if names:
-        compact["auth_surface_names_present"] = names
-    materialization = _compact_benchmark_post_launch_materialization(
-        value.get("post_launch_materialization")
-    )
-    if materialization:
-        compact["post_launch_materialization"] = materialization
-    closeout = value.get("closeout_command_templates")
-    if isinstance(closeout, dict):
-        compact_closeout: dict[str, Any] = {}
-        for field in (
-            "schema_version",
-            "display_command",
-            "post_run_rule",
-        ):
-            text = public_safe_compact_text(closeout.get(field), limit=320)
-            if text:
-                compact_closeout[field] = text
-        for field in (
-            "history_append",
-            "run_ledger_update",
-            "atomic_ledger_upsert",
-            "raw_paths_recorded",
-            "raw_logs_read",
-            "raw_task_text_read",
-        ):
-            if isinstance(closeout.get(field), bool):
-                compact_closeout[field] = closeout[field]
-        argv_template = public_safe_compact_list(
-            closeout.get("argv_template"),
-            limit=MAX_BENCHMARK_RUN_LIST_ITEMS * 4,
-        )
-        if argv_template:
-            compact_closeout["argv_template"] = argv_template
-        if compact_closeout:
-            compact["closeout_command_templates"] = compact_closeout
-    return compact
-
-
-def _compact_benchmark_repair_profile(value: Any) -> dict[str, Any]:
-    if not isinstance(value, dict):
-        return {}
-
-    compact: dict[str, Any] = {}
-    for field in ("schema_version", "repair_class"):
-        text = public_safe_compact_text(value.get(field), limit=140)
-        if text:
-            compact[field] = text
-    for field in (
-        "enabled",
-        "rerun_allowed_after_profile_applied",
-        "raw_logs_required",
-        "raw_task_text_required",
-        "credential_values_recorded",
-    ):
-        if isinstance(value.get(field), bool):
-            compact[field] = value[field]
-
-    for source_field in (
-        "required_launch_overrides",
-        "disallowed_launch_overrides",
-    ):
-        source = value.get(source_field)
-        if not isinstance(source, dict):
-            continue
-        compact_source: dict[str, Any] = {}
-        for key, raw_value in source.items():
-            safe_key = public_safe_compact_text(key, limit=100)
-            if not safe_key:
-                continue
-            if isinstance(raw_value, str):
-                text_value = public_safe_compact_text(raw_value, limit=140)
-                if text_value:
-                    compact_source[safe_key] = text_value
-            elif isinstance(raw_value, (int, float)) and not isinstance(raw_value, bool):
-                compact_source[safe_key] = raw_value
-            elif isinstance(raw_value, bool):
-                compact_source[safe_key] = raw_value
-        if compact_source:
-            compact[source_field] = compact_source
-    return compact
-
-
-def _compact_agent_setup_readiness(value: Any) -> dict[str, Any]:
-    if not isinstance(value, dict):
-        return {}
-
-    compact: dict[str, Any] = {}
-    for field in (
-        "schema_version",
-        "codex_install_strategy",
-        "first_blocker",
-        "next_action_after_setup_timeout",
-    ):
-        text = public_safe_compact_text(value.get(field), limit=180)
-        if text:
-            compact[field] = text
-    for field in (
-        "managed_codex_agent",
-        "worker_bridge_requested",
-        "worker_materialization_probe_only",
-        "runtime_codex_install_allowed",
-        "fail_fast_install_strategy",
-        "setup_timeout_budget_explicit",
-        "setup_timeout_repair_profile",
-        "same_task_repeat_after_setup_timeout_allowed",
-        "setup_failure_before_worker_counts_as_case_progress",
-        "raw_argv_recorded",
-        "raw_env_recorded",
-        "raw_logs_read",
-        "task_text_read",
-        "trajectory_read",
-        "credential_values_recorded",
-    ):
-        if isinstance(value.get(field), bool):
-            compact[field] = value[field]
-    multiplier = value.get("agent_setup_timeout_multiplier")
-    if isinstance(multiplier, (int, float)) and not isinstance(multiplier, bool):
-        compact["agent_setup_timeout_multiplier"] = multiplier
-    return compact
-
-
-def _compact_benchmark_post_launch_materialization(value: Any) -> dict[str, Any]:
-    if not isinstance(value, dict):
-        return {}
-
-    compact: dict[str, Any] = {}
-    for field in (
-        "schema_version",
-        "first_blocker",
-        "job_name",
-        "external_handle_kind",
-        "external_handle_state",
-        "compact_monitor_class",
-        "compact_failure_class",
-    ):
-        text = public_safe_compact_text(value.get(field), limit=140)
-        if text:
-            compact[field] = text
-    for field in (
-        "checked",
-        "ready_for_launch_state",
-        "ready_for_compact_result_ingest",
-        "jobs_dir_present",
-        "job_root_present",
-        "job_lock_present",
-        "job_result_present",
-        "ready_for_compact_failure_marker",
-        "external_handle_observed",
-        "external_handle_terminal",
-        "job_result_finished",
-        "job_active_without_trial_result",
-        "job_stale_active_without_trial_result",
-        "job_result_updated_at_present",
-        "stale_active_reconcile_requested",
-        "raw_paths_recorded",
-        "raw_logs_read",
-        "raw_task_text_read",
-        "trajectory_read",
-        "raw_external_handle_payload_recorded",
-    ):
-        if isinstance(value.get(field), bool):
-            compact[field] = value[field]
-    for field in (
-        "trial_result_present_count",
-        "candidate_job_root_count",
-        "job_running_trial_count",
-        "job_pending_trial_count",
-        "job_active_stale_seconds_threshold",
-    ):
-        count = value.get(field)
-        if isinstance(count, int) and not isinstance(count, bool):
-            compact[field] = count
-    age = value.get("job_updated_age_seconds")
-    if isinstance(age, (int, float)) and not isinstance(age, bool):
-        compact["job_updated_age_seconds"] = round(float(age), 3)
-    marker = value.get("compact_failure_marker")
-    if isinstance(marker, dict):
-        compact_marker: dict[str, Any] = {}
-        for field in (
-            "schema_version",
-            "failure_class",
-            "evidence_kind",
-            "external_handle_kind",
-            "external_handle_state",
-            "terminal_state",
-            "lifecycle_stage",
-            "ledger_attempt_kind",
-            "next_allowed_action",
-        ):
-            text = public_safe_compact_text(marker.get(field), limit=140)
-            if text:
-                compact_marker[field] = text
-        for field in (
-            "external_handle_terminal",
-            "terminal_closeout",
-            "runner_attempt_countable",
-            "launch_state_countable",
-            "case_attempt_countable",
-            "benchmark_budget_countable",
-            "job_result_present",
-            "job_result_finished",
-            "job_result_updated_at_present",
-            "raw_paths_recorded",
-            "raw_logs_read",
-            "raw_task_text_read",
-            "trajectory_read",
-            "raw_external_handle_payload_recorded",
-        ):
-            if isinstance(marker.get(field), bool):
-                compact_marker[field] = marker[field]
-        trial_result_count = marker.get("trial_result_present_count")
-        if isinstance(trial_result_count, int) and not isinstance(
-            trial_result_count, bool
-        ):
-            compact_marker["trial_result_present_count"] = trial_result_count
-        for field in (
-            "job_running_trial_count",
-            "job_pending_trial_count",
-            "job_active_stale_seconds_threshold",
-        ):
-            count = marker.get(field)
-            if isinstance(count, int) and not isinstance(count, bool):
-                compact_marker[field] = count
-        age = marker.get("job_updated_age_seconds")
-        if isinstance(age, (int, float)) and not isinstance(age, bool):
-            compact_marker["job_updated_age_seconds"] = round(float(age), 3)
-        attempt_accounting = _compact_benchmark_attempt_accounting(
-            marker.get("attempt_accounting")
-        )
-        if attempt_accounting:
-            compact_marker["attempt_accounting"] = attempt_accounting
-        if compact_marker:
-            compact["compact_failure_marker"] = compact_marker
-    return compact
-
-
-def compact_benchmark_post_launch_materialization(value: Any) -> dict[str, Any] | None:
-    compact = _compact_benchmark_post_launch_materialization(value)
-    return compact or None
-
-
-def compact_benchmark_run(run: dict[str, Any]) -> dict[str, Any] | None:
-    source = _benchmark_run_source_read_model(
-        run,
-        schema_version=BENCHMARK_RUN_SCHEMA_VERSION,
-    )
-    if not source:
-        return None
-
-    trials_source = (
-        source.get("trials") if isinstance(source.get("trials"), list) else []
-    )
-    compact = _compact_benchmark_run_core_read_model(
-        source,
-        schema_version=BENCHMARK_RUN_SCHEMA_VERSION,
-        max_list_items=MAX_BENCHMARK_RUN_LIST_ITEMS,
-    )
-    execution_contract = _compact_benchmark_run_execution_contract(
-        source,
-        max_list_items=MAX_BENCHMARK_RUN_LIST_ITEMS,
-    )
-    loop_contract = execution_contract.get("benchmark_loop_contract")
-    if loop_contract:
-        compact["benchmark_loop_contract"] = loop_contract
-    compact.update(
-        _compact_benchmark_run_pre_execution_metadata(
-            source,
-            max_list_items=MAX_BENCHMARK_RUN_LIST_ITEMS,
-        )
-    )
-
-    compact.update(
-        {
-            field: execution_contract[field]
-            for field in ("claim_boundary", "agent", "model_control")
-            if field in execution_contract
-        }
-    )
-
-    post_execution = _compact_benchmark_run_post_execution_metadata(
-        source,
-        max_list_items=MAX_BENCHMARK_RUN_LIST_ITEMS,
-    )
-    compact.update(post_execution.failure)
-
-    compose_setup_diagnostic = _compact_benchmark_compose_setup_diagnostic(
-        source.get("compose_setup_diagnostic")
-    )
-    if compose_setup_diagnostic:
-        compact["compose_setup_diagnostic"] = compose_setup_diagnostic
-        _apply_skillsbench_pre_agent_setup_compact_projection(compact)
-
-    compact.update(post_execution.progress_metrics)
-
-    interaction_counters = _compact_benchmark_interaction_counters(
-        source.get("interaction_counters")
-    )
-    if interaction_counters:
-        compact["interaction_counters"] = interaction_counters
-
-    compact.update(post_execution.lifecycle)
-    _repair_product_mode_lifecycle_missing_attribution_read_model(
-        compact,
-        max_list_items=MAX_BENCHMARK_RUN_LIST_ITEMS,
-    )
-
-    preflight_guard = _compact_benchmark_preflight_guard(source.get("preflight_guard"))
-    if preflight_guard:
-        compact["preflight_guard"] = preflight_guard
-
-    compact.update(post_execution.active_user_and_claim)
-
-    private_runner_launch = _compact_benchmark_private_runner_launch(
-        source.get("private_runner_launch_summary")
-    )
-    if private_runner_launch:
-        compact["private_runner_launch_summary"] = private_runner_launch
-
-    setup_timeout_repair_profile = _compact_benchmark_repair_profile(
-        source.get("setup_timeout_repair_profile")
-    )
-    if setup_timeout_repair_profile:
-        compact["setup_timeout_repair_profile"] = setup_timeout_repair_profile
-
-    compact.update(post_execution.worker_outcome)
-
-    compact_validation = _compact_benchmark_run_validation_read_model(
-        source.get("validation"),
-        pre_agent_setup_materialization_blocked=(
-            compact.get("pre_agent_setup_materialization_blocked") is True
-        ),
-        max_list_items=MAX_BENCHMARK_RUN_LIST_ITEMS,
-    )
-    if compact_validation:
-        compact["validation"] = compact_validation
-        _apply_skillsbench_pre_agent_setup_compact_projection(compact)
-
-    _apply_skillsbench_runner_source_fingerprint_compact_projection(
-        compact,
-        source=source,
-    )
-    _apply_skillsbench_benchmark_egress_preflight_compact_projection(
-        compact,
-        source=source,
-    )
-    apply_skillsbench_verifier_bootstrap_missing_score_attribution(
-        compact,
-        task_staging=compact.get("task_staging"),
-        setup_preflight=compact.get("task_setup_preflight"),
-    )
-
-    _sync_skillsbench_runner_failure_root_blockers(compact)
-
-    if solution_quality := build_benchmark_solution_quality_signals(compact):
-        compact["solution_quality_signals"] = solution_quality
-
-    post_run_debug_gate = build_skillsbench_post_run_debug_gate(compact)
-    if post_run_debug_gate:
-        compact["post_run_debug_gate"] = post_run_debug_gate
-
-    trials = _compact_benchmark_run_trials_read_model(
-        trials_source,
-        max_trials=MAX_BENCHMARK_RUN_TRIALS,
-        max_list_items=MAX_BENCHMARK_RUN_LIST_ITEMS,
-    )
-    if trials:
-        compact["trials"] = trials
-        raw_trials = source.get("trials")
-        if isinstance(raw_trials, list):
-            compact["trial_count"] = len(raw_trials)
-
-    for field in ("evidence_files", "resume_or_inspect_commands", "stop_conditions"):
-        values = public_safe_compact_list(source.get(field), limit=MAX_BENCHMARK_RUN_LIST_ITEMS)
-        if values:
-            compact[field] = values
-
-    read_boundary = source.get("read_boundary")
-    if isinstance(read_boundary, dict):
-        compact_boundary: dict[str, bool] = {}
-        for field in (
-            "compact_only",
-            "raw_artifacts_read",
-            "task_text_read",
-            "trajectory_read",
-            "controller_trace_read",
-            "local_paths_recorded",
-            "docker_invoked",
-            "model_api_invoked",
-            "upload_invoked",
-        ):
-            if isinstance(read_boundary.get(field), bool):
-                compact_boundary[field] = read_boundary[field]
-        if compact_boundary:
-            compact["read_boundary"] = compact_boundary
-
-    if set(compact.keys()) == {"schema_version"}:
-        return None
-    return compact
 
 
 def state_event_log_candidates(goal: dict[str, Any], *, state_path: Path) -> list[Path]:
@@ -2781,6 +1225,30 @@ def build_attention_queue(
         include_task_graph=include_task_graph,
         goal_id_filter=goal_id_filter,
     )
+
+
+def _compact_benchmark_post_launch_materialization(value: Any) -> dict[str, Any] | None:
+    from .benchmarks.read_models.benchmark_status_runner import (
+        _compact_benchmark_post_launch_materialization as _impl,
+    )
+
+    return _impl(value)
+
+
+def compact_benchmark_post_launch_materialization(value: Any) -> dict[str, Any] | None:
+    from .benchmarks.read_models.benchmark_status_runner import (
+        compact_benchmark_post_launch_materialization as _impl,
+    )
+
+    return _impl(value)
+
+
+def compact_benchmark_run(run: dict[str, Any]) -> dict[str, Any] | None:
+    from .benchmarks.read_models.benchmark_status_runner import (
+        compact_benchmark_run as _impl,
+    )
+
+    return _impl(run)
 
 
 def compact_run(run: dict[str, Any]) -> dict[str, Any]:
